refactor: rename client files
