+++ src/client/es/hmr-client-content.ts
@@ -9,7 +9,7 @@ import type {
   UpdatePayload,
 } from 'vite'
 
-import { ErrorOverlay, overlayId } from './content-hmr-overlay'
+import { ErrorOverlay, overlayId } from './hmr-client-content-overlay'
 
 // See https://stackoverflow.com/a/63549561.
 type CustomEventName<T extends string> = (T extends `vite:${T}` ? never : T) &
